% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{alon_data}
\alias{alon_data}
\title{Normalized gene expression values from Alon et al. (1999).}
\format{A data frame with 2000 rows (genes) and 62 variables (samples).}
\usage{
data(alon_data)
}
\description{
A dataset containing centred and normalized values of the 
logged expression values of a subset of 2000 genes taken from 
Alon, Uri, et al. "Broad patterns of gene expression revealed by clustering
analysis of tumor and normal colon tissues probed by oligonucleotide arrays."
Proceedings of the National Academy of Sciences 96.12 (1999): 6745-6750.
The method of subset selection was described in G. J. McLachlan, R. W. Bean, 
D. Peel; A mixture model-based approach to the clustering of microarray 
expression data ,
Bioinformatics, Volume 18, Issue 3, 1 March 2002, Pages 413–422.
}
\examples{
dim(alon_data)
}
\keyword{datasets}
